Conflicts:
	phprojekt/application/Calendar2/Views/dojo/scripts/Main.js
	phprojekt/htdocs/dojo/util/buildscripts/profiles/phprojekt.profile.js

+++ phprojekt/application/Default/Controllers/IndexController.php
@@ -18,7 +18,7 @@
  * @license    LGPL v3 (See LICENSE file)
  * @link       http://www.phprojekt.com
  * @since      File available since Release 6.0
- * @version    Release: @package_version@
+ * @version    Release: 6.1.0
  * @author     David Soria Parra <soria_parra@mayflower.de>
  */
 
@@ -39,7 +39,7 @@
  * @license    LGPL v3 (See LICENSE file)
  * @link       http://www.phprojekt.com
  * @since      File available since Release 6.0
- * @version    Release: @package_version@
+ * @version    Release: 6.1.0
  * @author     David Soria Parra <soria_parra@mayflower.de>
  */
 class IndexController extends Zend_Controller_Action
@@ -315,8 +315,7 @@ class IndexController extends Zend_Controller_Action
         $moduleName = $this->getModuleName();
         $object     = Phprojekt_Loader::getModel($modelName, $moduleName);
         if (null === $object) {
-            /* @todo throw error */
-            $object = new Default_Models_Default();
+            throw new Exception('No Model object could be found');
         }
 
         return $object;
@@ -707,6 +706,12 @@ class IndexController extends Zend_Controller_Action
         }
 
         $model = $this->getModelObject()->find($id);
+        if (empty($model)) {
+            throw new Zend_Controller_Action_Exception(self::NOT_FOUND, 404);
+        }
+        if ($model->hasField('projectId')) {
+            Phprojekt::setCurrentProjectId($model->projectId);
+        }
 
         if ($model instanceof Phprojekt_ActiveRecord_Abstract) {
             $tmp = Default_Helpers_Delete::delete($model);
@@ -750,6 +755,7 @@ class IndexController extends Zend_Controller_Action
     public function jsonDeleteMultipleAction()
     {
         $ids = $this->getRequest()->getParam('ids');
+        $this->setCurrentProjectId();
 
         if (!empty($ids)) {
             $message  = Phprojekt::getInstance()->translate(self::DELETE_MULTIPLE_TRUE_TEXT);
@@ -1183,40 +1189,11 @@ class IndexController extends Zend_Controller_Action
      */
     public function fileDownloadAction()
     {
-        $order = (int) $this->getRequest()->getParam('order', null);
+        $hash = (int) $this->getRequest()->getParam('hash', null);
 
         list($model, $field, $itemId) = $this->_getFileParameters();
 
-        Default_Helpers_Upload::downloadFile($model, $field, $itemId, $order);
-    }
-
-    /**
-     * Deletes a file and call the render for draw the upload field.
-     *
-     * OPTIONAL request parameters:
-     * <pre>
-     *  - integer <b>id</b>    Id of the current item.
-     *  - string  <b>field</b> Name of the field in the module.
-     *  - integer <b>order</b> Position of the file (Can be many uploaded files in the same field).
-     * </pre>
-     *
-     * @return void
-     */
-    public function fileDeleteAction()
-    {
-        $order = (int) $this->getRequest()->getParam('order', 0);
-
-        list($model, $field, $itemId) = $this->_getFileParameters();
-
-        try {
-            $value = Default_Helpers_Upload::deleteFile($model, $field, $itemId, $order);
-        } catch (Exception $error) {
-            // Show error to user
-            $this->view->errorMessage = $error->getMessage();
-            $value                    = Default_Helpers_Upload::getFiles($model, $field);
-        }
-
-        $this->_fileRenderView($itemId, $field, $value);
+        Default_Helpers_Upload::downloadFile($model, $field, $itemId, $hash);
     }
 
     /**
@@ -1246,7 +1223,7 @@ class IndexController extends Zend_Controller_Action
      *
      * @return void
      */
-    private function _fileRenderView($itemId, $field, $value)
+    private function _fileRenderView($itemId, $field, $files)
     {
         $this->getResponse()->clearHeaders();
         $this->getResponse()->clearBody();
@@ -1255,6 +1232,7 @@ class IndexController extends Zend_Controller_Action
         $csrfNamespace = new Zend_Session_Namespace($sessionName);
         $config        = Phprojekt::getInstance()->getConfig();
         $linkBegin     = $config->webpath . 'index.php/' . $this->getModuleName() . '/index/';
+        $fieldId       = $this->getRequest()->getParam('fieldId', '');
 
         // Add all the extra parameters that have the original URL
         $linkData      = '';
@@ -1273,31 +1251,42 @@ class IndexController extends Zend_Controller_Action
         $this->view->fileName       = null;
         $this->view->itemId         = $itemId;
         $this->view->field          = $field;
-        $this->view->value          = $value;
+        $this->view->fieldId        = $fieldId;
         $this->view->csrfToken      = $csrfNamespace->token;
         $this->view->maxUploadSize  = (isset($config->maxUploadSize)) ? (int) $config->maxUploadSize :
             Phprojekt::DEFAULT_MAX_UPLOAD_SIZE;
 
-        $filesForView = array();
+        $model = $this->getModelObject();
+        $model->find($itemId);
+
+        $filesForView         = array();
+        $hasDownloadRight     = $model->hasRight(Phprojekt_Auth_Proxy::getEffectiveUserId(), Phprojekt_Acl::DOWNLOAD);
+        $hasWriteRight        = $model->hasRight(Phprojekt_Auth_Proxy::getEffectiveUserId(), Phprojekt_Acl::WRITE);
+        $this->view->disabled = !$hasWriteRight;
 
         // Is there any file?
-        if (!empty($value)) {
-            $files = explode('||', $value);
-            $model = $this->getModelObject();
-            $model->find($itemId);
-            $i      = 0;
-            foreach ($files as $file) {
-                $fileName = strstr($file, '|');
-                $fileData = 'id/' . $itemId . '/field/' . $field . '/order/'
-                    . (string) ($i + 1) . '/csrfToken/' . $csrfNamespace->token;
+        if (!empty($files)) {
+            $i = 0;
 
-                $filesForView[$i] = array('fileName' => substr($fileName, 1));
-                if ($model->hasRight(Phprojekt_Auth_Proxy::getEffectiveUserId(), Phprojekt_Acl::DOWNLOAD)) {
+            foreach ($files as $file) {
+                $fileName = $file['name'];
+                $fileHash = $file['md5'];
+                $fileData = 'id/' . $itemId . '/field/' . $field . '/hash/' . $fileHash . '/csrfToken/' . $csrfNamespace->token;
+
+                $filesForView[$i] = array(
+                    'fileName' => $fileName,
+                    'hash' => $fileHash
+                );
+
+                if ($hasDownloadRight) {
                     $filesForView[$i]['downloadLink'] = $linkBegin . 'fileDownload/' . $linkData . $fileData;
                 }
-                if ($model->hasRight(Phprojekt_Auth_Proxy::getEffectiveUserId(), Phprojekt_Acl::WRITE)) {
-                    $filesForView[$i]['deleteLink'] = $linkBegin . 'fileDelete/' . $linkData . $fileData;
-                }
+
+                $fileinfo = Default_Helpers_Upload::getInfosFromFile($file);
+
+                $filesForView[$i]['size'] = $fileinfo['size'];
+                $filesForView[$i]['ctime'] = $fileinfo['ctime'];
+
                 $i++;
             }
         }
